requirements.txt
fastapi
uvicorn
python-dotenv
autogen-agentchat
python-multipart
pydantic
flaml[automl]
requests